Refact myposts.vue - move the favorite page logic into a composable
